feat: added integration with d3fc-label-layout
BREAKING CHANGE: this removes the 'rectangles' layout component replacing it with the equivalent 'label' layout component from d3fc-label-layout
