+++ site/src/components/layout/rectangles-example-measure.html
@@ -1,1 +1,1 @@
-<div id="rectangles-measure" class='chart large'></div>
+<div id="label" class='chart large'></div>
