+++ src/layout/layout.js
@@ -1,7 +1,13 @@
-import rectangles from './rectangles';
-import strategy from './strategy/strategy';
+import label from './label';
+import { boundingBox, greedy, annealing, removeOverlaps, textLabel } from 'd3fc-label-layout';
 
 export default {
-    rectangles: rectangles,
-    strategy: strategy
+    label: label,
+    textLabel: textLabel,
+    strategy: {
+        boundingBox: boundingBox,
+        greedy: greedy,
+        annealing: annealing,
+        removeOverlaps: removeOverlaps
+    }
 };
