--- src/layout/strategy/strategy.js
@@ -1,13 +1,0 @@
-import boundingBox from './boundingBox';
-import greedy from './greedy';
-import local from './local';
-import annealing from './annealing';
-import removeOverlaps from './removeOverlaps';
-
-export default {
-    boundingBox: boundingBox,
-    greedy: greedy,
-    local: local,
-    annealing: annealing,
-    removeOverlaps: removeOverlaps
-};
