+++ visual-tests/index.html
@@ -284,11 +284,6 @@
               <p>Demonstrates the Layout component, applies the CSS flexbox layout to SVG in order to simplify the task of constructing charts.</p>
             </article>
 
-            <article data-href="layout/rectangles.html">
-              <h3>Rectangles</h3>
-              <p>Demonstrates the rectangle layout component.</p>
-            </article>
-
             <article data-href="layout/layoutProgrammatic.html">
               <h3>Layout Programmatic</h3>
               <p>Constructs a layout programmatically.</p>
