--- visual-tests/layout/rectangles.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Smart Container - d3fc Visual Tests</title>
-    <link href="../assets/d3fc.css" rel="stylesheet">
-  </head>
-  <body>
-    <div id="tooltip" style="float: left;"></div>
-    <div id="strategies" style="float: right;">
-      Use Strategy:
-      <p>
-        <input type="radio" id="noop" name="strategy" value="noop" checked>
-        <label for="noop">No-op</label>
-      </p>
-      <p>
-        <input type="radio" id="bounding-box" name="strategy" value="bounding-box">
-        <label for="bounding-box">Bounding Box</label>
-      </p>
-      <p>
-        <input type="radio" id="greedy" name="strategy" value="greedy">
-        <label for="greedy">Greedy</label>
-      </p>
-      <p>
-        <input type="radio" id="local" name="strategy" value="local">
-        <label for="local">Local</label>
-      </p>
-      <p>
-        <input type="radio" id="annealing" name="strategy" value="annealing">
-        <label for="annealing">Annealing</label>
-      </p>
-    </div>
-    <footer>
-      <script src="../assets/d3.js"></script>
-      <script src="../assets/css-layout.js"></script>
-      <script src="../assets/d3fc.js"></script>
-      <script src="../assets/seedrandom.min.js"></script>
-      <script src="../setup.js"></script>
-      <script src="rectangles.js"></script>
-      <script src="//localhost:36729/livereload.js"></script>
-    </footer>
-  </body>
-</html>
